chore(team): split task manager tests
